Extract build logic from main build script to plugins
