+++ junit-jupiter-api/junit-jupiter-api.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
-	`java-library`
-	kotlin("jvm")
+	`kotlin-library-conventions`
 }
 
 description = "JUnit Jupiter API"
