+++ junit-jupiter-engine/junit-jupiter-engine.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
-	`java-library`
-	kotlin("jvm")
+	`kotlin-library-conventions`
 	groovy
 }
