+++ junit-jupiter/junit-jupiter.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-	`java-library`
+	`java-library-conventions`
 }
 
 description = "JUnit Jupiter (Aggregator)"
