+++ junit-platform-suite-api/junit-platform-suite-api.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-	`java-library`
+	`java-library-conventions`
 }
 
 description = "JUnit Platform Suite API"
