+++ junit-platform-testkit/junit-platform-testkit.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-	`java-library`
+	`java-library-conventions`
 }
 
 description = "JUnit Platform Test Kit"
