+++ platform-tests/platform-tests.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-	`java-library`
+	`java-library-conventions`
 	id("me.champeau.gradle.jmh")
 }
